+++ src/modules/admin/dto/response/reportDetailRes.dto.ts
@@ -1,8 +1,8 @@
-import { ReportDto } from './report.dto';
+import { ReportDto } from '../report.dto';
 import { ApiProperty } from '@nestjs/swagger';
 import { Expose } from 'class-transformer';
 import { IsBoolean, IsDate, IsNumber, IsString } from 'class-validator';
-import { ReportsDto } from './reports.dto';
+import { ReportsDto } from '../reports.dto';
 
 export class ReportDetailResDto {
   @ApiProperty()
