+++ src/modules/admin/dto/response/reviewRes.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Expose } from 'class-transformer';
-import { ReviewsDto } from './reviews.dto';
+import { ReviewsDto } from '../reviews.dto';
 
 export class ReviewResDto {
   @ApiProperty({ type: [ReviewsDto] })
