+++ src/modules/auth/dto/response/userRes.dto.ts
@@ -1,12 +1,17 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Exclude } from 'class-transformer';
-import { IsDate, IsEmail, IsNotEmpty, IsNumber, IsString } from 'class-validator';
+import { IsDate, IsEmail, IsNotEmpty, IsNumber, IsOptional, IsString } from 'class-validator';
 
 export class UserResDto {
   @ApiProperty()
   @IsNotEmpty()
   @IsNumber()
   id: number;
+
+  @ApiProperty()
+  @IsOptional()
+  @IsString()
+  nickname: string;
 
   @ApiProperty()
   @IsString()
