+++ src/modules/auth/tests/auth.service.spec.ts
@@ -2,8 +2,8 @@ import { AuthRepository } from '../auth.repository';
 import { AuthService } from '../auth.service';
 import { Test, TestingModule } from '@nestjs/testing';
 import { HttpException } from '@nestjs/common';
-import { CheckEmailReqDto } from '../dto/checkEamilReq.dto';
-import { CreateUserReqDto } from '../dto/createUserReq.dto';
+import { CheckEmailReqDto } from '../dto/request/checkEamilReq.dto';
+import { CreateUserReqDto } from '../dto/request/createUserReq.dto';
 import { User } from '../../../entities/user.entity';
 import { MailService } from '../../mail/mail.service';
 import * as bcrypt from 'bcrypt';
